+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - More aliases for many wwctl commands
 - Add support to render template using `host` or `$(uname -n)` as the value of `overlay show --render`. #623
 - Added command line parameters for credentials of a container registry
+- Add `wwctl clean` to remove OCI cache and overlays from deleted nodes
 
 ### Changed